Changes to options, transitions, rooms, healths and mores
A lot of changes. That included adding option to fight or flee and rest or continue forward. Added transitions to room 1 to 4. Enemy in each room have a bit of personalty. Health is recover every time you enter a new room (Probably change this in the future), and room 4 is the end for now.